# autenticacion/src/test/resources/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Deployed Resources",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8090/",
      "description": "Default Server."
    }
  ],
  "paths": {
...
    "/api/usuarios": {
      "get": {
        "summary": "lista los usuarios",
        "description": "Permite obtener un listado de los usuarios registrados en el sistema, es posible filtrarlos por rol y por el nombre de usuario, de igual forma es posible ordenarlos ASC o DESC",
        "operationId": "UsuarioAPI :: listar usuario",
        "parameters": [
          {
            "name": "usuario",
            "in": "query",
            "description": "nombre del usuario por el que se desea filtrar",
            "required": false,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "pedro"
          },
          {
            "name": "rol",
            "in": "query",
            "description": "rol por el que se desea filtrar, si es especificado solo se mostrarÃ¡n aquellos usuarios que posean dicho rol.",
            "required": false,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "admin"
          },
          {
            "name": "order",
            "in": "query",
            "description": "El listado es ordenado por nombre de usuario, order permite indicar si se desea en orden ascendente (ASC) o descendente (DESC).",
            "required": false,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "DESC"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuarios"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Usuario no autorizado\"}]"
              }
            },
            "description": "Usuario no autorizado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"No tiene permitido realizar la operaciÃ³n\"}]"
              }
            },
            "description": "No tiene permitido realizar la operaciÃ³n"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Recurso no encontrado\"}]"
              }
            },
            "description": "Recurso no encontrado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false,
        "security": [
          {
            "JWT": [
              "admin"
            ]
          }
        ]
      },
      "post": {
        "summary": "Registrar usuario",
        "description": "Permite registrar un usuario en el sistema",
        "operationId": "UsuarioAPI :: registrar",
        "requestBody": {
          "description": "Usuario a registrar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"El usuario y la contraseÃ±a son requeridos\"}]"
              }
            },
            "description": "La solicitud estÃ¡ incompleta"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"El usuario ya existe\"}]"
              }
            },
            "description": "El usuario ya existe"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false
      }
    },
    "/api/usuarios/{username}": {
      "get": {
        "summary": "obtener datos de usuario",
        "description": "Permite obtener los datos de un usuario",
        "operationId": "UsuarioAPI :: consultar usuario",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "nombre del usuario del que se desea obtener los datos",
            "required": true,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "pedro"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Usuario no autorizado\"}]"
              }
            },
            "description": "Usuario no autorizado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"No tiene permitido realizar la operaciÃ³n\"}]"
              }
            },
            "description": "No tiene permitido realizar la operaciÃ³n"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Recurso no encontrado\"}]"
              }
            },
            "description": "Recurso no encontrado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false,
        "security": [
          {
            "JWT": [
              "user",
              "admin"
            ]
          }
        ]
      },
      "put": {
        "summary": "Actualizar usuario",
        "description": "Permite actualizar los datos de un usuario en el sistema",
        "operationId": "UsuarioAPI :: actualizar",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "nombre del usuario que se desea actualizar",
            "required": true,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "pedro"
          }
        ],
        "requestBody": {
          "description": "Datos del usuario a actualizar",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            },
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                }
              }
            },
            "description": "Datos incompletos"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Usuario no autorizado\"}]"
              }
            },
            "description": "Usuario no autorizado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"No tiene permitido realizar la operaciÃ³n\"}]"
              }
            },
            "description": "No tiene permitido realizar la operaciÃ³n"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Recurso no encontrado\"}]"
              }
            },
            "description": "Recurso no encontrado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false,
        "security": [
          {
            "JWT": [
              "user",
              "admin"
            ]
          }
        ]
      },
      "delete": {
        "summary": "Dar de baja un usuario",
        "description": "Permite dar de baja (borrar) un usuario en el sistema",
        "operationId": "UsuarioAPI :: unregister",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "nombre del usuario que se desea dar de baja",
            "required": true,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "pedro"
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Usuario no autorizado\"}]"
              }
            },
            "description": "Usuario no autorizado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"No tiene permitido realizar la operaciÃ³n\"}]"
              }
            },
            "description": "No tiene permitido realizar la operaciÃ³n"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Recurso no encontrado\"}]"
              }
            },
            "description": "Recurso no encontrado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false,
        "security": [
          {
            "JWT": [
              "user",
              "admin"
            ]
          }
        ]
      },
      "patch": {
        "summary": "Actualizar clave",
        "description": "Permite actualizar la clave de un usuario en el sistema",
        "operationId": "UsuarioAPI :: actualizar clave",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "nombre del usuario al que se desea actualizar la clave",
            "required": true,
            "deprecated": false,
            "allowEmptyValue": false,
            "allowReserved": false,
            "schema": {
              "type": "string"
            },
            "example": "pedro"
          }
        ],
        "requestBody": {
          "description": "Claves necesarias para la actualizaciÃ³n",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Clave"
              }
            },
            "*/*": {
              "schema": {
                "$ref": "#/components/schemas/Clave"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usuario"
                }
              }
            },
            "description": "OperaciÃ³n exitosa"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                }
              }
            },
            "description": "La solicitud estÃ¡ incompleta, el clave actual, la nueva clave y la verificaciÃ³n de la nueva clave son requeridas"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Usuario no autorizado\"}]"
              }
            },
            "description": "Usuario no autorizado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"No tiene permitido realizar la operaciÃ³n\"}]"
              }
            },
            "description": "No tiene permitido realizar la operaciÃ³n"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"Recurso no encontrado\"}]"
              }
            },
            "description": "Recurso no encontrado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Errores"
                },
                "example": "[{\"error\":\"OcurriÃ³ un error interno al procesar la solicitud\"}]"
              }
            },
            "description": "Error interno"
          }
        },
        "deprecated": false,
        "security": [
          {
            "JWT": [
              "user",
              "admin"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "Clave": {
        "exclusiveMaximum": false,
        "exclusiveMinimum": false,
        "minLength": 0,
        "uniqueItems": false,
        "maxProperties": 0,
        "minProperties": 0,
        "type": "object",
        "properties": {
          "clave": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "string",
            "description": "Nueva clave a ser asignada al usuario",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "deprecated": false
          },
          "nueva_clave": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "string",
            "description": "Nueva clave a ser asignada al usuario",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "deprecated": false
          },
          "verificacion_clave": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "string",
            "description": "Nueva clave repetida para verificaciÃ³n",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "deprecated": false
          }
        },
        "nullable": false,
        "readOnly": false,
        "writeOnly": false,
        "example": "{\"clave\":\"12345\",\"nueva_clave\":\"ABC123\",\"verificacion_clave\":\"ABC123\"}",
        "deprecated": false
      },
...
      "Errores": {
        "exclusiveMaximum": false,
        "exclusiveMinimum": false,
        "minLength": 0,
        "uniqueItems": false,
        "maxProperties": 0,
        "minProperties": 0,
        "type": "array",
        "nullable": false,
        "readOnly": false,
        "writeOnly": false,
        "example": "[{\"error\":\"Se produjo un error al procesar la solicitud\"}]",
        "deprecated": false,
        "items": {
          "exclusiveMaximum": false,
          "exclusiveMinimum": false,
          "minLength": 0,
          "uniqueItems": false,
          "maxProperties": 0,
          "minProperties": 0,
          "type": "object",
          "properties": {
            "error": {
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "minLength": 0,
              "uniqueItems": false,
              "maxProperties": 0,
              "minProperties": 0,
              "type": "string",
              "description": "DescripciÃ³n del error ocurrido",
              "nullable": false,
              "readOnly": false,
              "writeOnly": false,
              "deprecated": false
            }
          },
          "nullable": false,
          "readOnly": false,
          "writeOnly": false,
          "example": "{\"error\":\"Se produjo un error al procesar la solicitud\"}",
          "deprecated": false
        }
      },
      "Message": {
        "exclusiveMaximum": false,
        "exclusiveMinimum": false,
        "minLength": 0,
        "uniqueItems": false,
        "maxProperties": 0,
        "minProperties": 0,
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "nullable": false,
        "readOnly": false,
        "writeOnly": false,
        "example": "{\"error\":\"Se produjo un error al procesar la solicitud\"}",
        "deprecated": false
      },
...
      "Usuario": {
        "exclusiveMaximum": false,
        "exclusiveMinimum": false,
        "minLength": 0,
        "uniqueItems": false,
        "maxProperties": 0,
        "minProperties": 0,
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "roles": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "array",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "example": "[\"user\"]",
            "deprecated": false,
            "items": {
              "type": "string"
            }
          },
          "usuario": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "string",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "example": "pedro",
            "deprecated": false
          },
          "clave": {
            "exclusiveMaximum": false,
            "exclusiveMinimum": false,
            "minLength": 0,
            "uniqueItems": false,
            "maxProperties": 0,
            "minProperties": 0,
            "type": "string",
            "nullable": false,
            "readOnly": false,
            "writeOnly": false,
            "example": "12345",
            "deprecated": false
          }
        },
        "nullable": false,
        "readOnly": false,
        "writeOnly": false,
        "example": "{\"usuario\":\"pedro\",\"clave\":\"12345\",\"roles\":[\"user\"]}",
        "deprecated": false
      },
      "Usuarios": {
        "exclusiveMaximum": false,
        "exclusiveMinimum": false,
        "minLength": 0,
        "uniqueItems": false,
        "maxProperties": 0,
        "minProperties": 0,
        "type": "array",
        "nullable": false,
        "readOnly": false,
        "writeOnly": false,
        "example": "[{\"usuario\":\"pedro\",\"clave\":\"12345\",\"roles\":[\"user\"]}]",
        "deprecated": false,
        "items": {
          "exclusiveMaximum": false,
          "exclusiveMinimum": false,
          "minLength": 0,
          "uniqueItems": false,
          "maxProperties": 0,
          "minProperties": 0,
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "roles": {
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "minLength": 0,
              "uniqueItems": false,
              "maxProperties": 0,
              "minProperties": 0,
              "type": "array",
              "nullable": false,
              "readOnly": false,
              "writeOnly": false,
              "example": "[\"user\"]",
              "deprecated": false,
              "items": {
                "type": "string"
              }
            },
            "usuario": {
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "minLength": 0,
              "uniqueItems": false,
              "maxProperties": 0,
              "minProperties": 0,
              "type": "string",
              "nullable": false,
              "readOnly": false,
              "writeOnly": false,
              "example": "pedro",
              "deprecated": false
            },
            "clave": {
              "exclusiveMaximum": false,
              "exclusiveMinimum": false,
              "minLength": 0,
              "uniqueItems": false,
              "maxProperties": 0,
              "minProperties": 0,
              "type": "string",
              "nullable": false,
              "readOnly": false,
              "writeOnly": false,
              "example": "12345",
              "deprecated": false
            }
          },
          "nullable": false,
          "readOnly": false,
          "writeOnly": false,
          "example": "{\"usuario\":\"pedro\",\"clave\":\"12345\",\"roles\":[\"user\"]}",
          "deprecated": false
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}